@@ -1,5 +1,5 @@
-name: flutterapp
-description: A new Flutter application.
+name: blogapp
+description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
@@ -12,13 +12,15 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # build by specifying --build-name and --build-number, respectively.
 # In Android, build-name is used as versionName while build-number used as versionCode.
 # Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
+# In iOS, build-name is used as CFBundleShortVersionString while build-number is used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
+# In Windows, build-name is used as the major, minor, and patch parts
+# of the product and file versions while build-number is used as the build suffix.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: '>=2.18.6 <3.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -27,18 +29,24 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
+  counter_button: ^1.0.1
   cupertino_icons: ^1.0.2
+  customizable_counter: ^1.0.4
   flutter:
     sdk: flutter
-  http: ^0.13.4
-  loading_indicator: ^3.1.0
-  shared_preferences: ^2.0.16
-
+  google_maps_flutter: ^2.2.3
+  http: ^0.13.3
+  image_picker: ^0.7.5+4
+  map_launcher: ^2.4.0
+  maps_launcher: ^2.0.1
+  shared_preferences: ^2.0.6
+  stepper_counter_swipe: ^1.0.2
+  url_launcher: ^6.1.9
+  flutter_svg: ^1.1.6
 dev_dependencies:
   flutter_launcher_icons: "^0.11.0"
   flutter_test:
     sdk: flutter
-
 flutter_icons:
   android: "launcher_icon"
   ios: true
@@ -49,11 +57,11 @@ flutter_icons:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0
+  flutter_lints: ^2.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
-# The following section is specific to Flutter.
+# The following section is specific to Flutter packages.
 flutter:
 
   # The following line ensures that the Material Icons font is
@@ -64,10 +72,9 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/imgs/
-  #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
   # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
+  # https://flutter.dev/assets-and-images/#resolution-aware
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
   # To add custom fonts to your application, add a fonts section here,
