@@ -1,11 +1,600 @@
 import 'package:flutter/material.dart';
+import 'package:blogapp/screens/product_detail.dart';
+import 'package:blogapp/screens/product_page.dart';
+import '../constant.dart';
 
 class BeerHouse extends StatelessWidget {
+  ProductPage productpage = ProductPage();
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: Text('Hola Mundo'),
+    return Scaffold(
+      backgroundColor: Color(0xFFFCFAF8),
+      body: ListView(
+        children: <Widget>[
+          SizedBox(height: 15.0),
+          Container(
+              padding: EdgeInsets.only(right: 15.0),
+              width: MediaQuery.of(context).size.width - 30.0,
+              height: MediaQuery.of(context).size.height - 50.0,
+              child: GridView.count(
+                crossAxisCount: 2,
+                primary: false,
+                crossAxisSpacing: 10.0,
+                mainAxisSpacing: 15.0,
+                childAspectRatio: 0.8,
+                children: <Widget>[
+                  productpage.build(
+                      'Barrilito 355 ml',
+                      '\$11.00',
+                      'imgs/Beerhouse/Barrilito355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Barrilito 444ml',
+                      '\$13.00',
+                      'imgs/Beerhouse/Barrilito444ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Bocanegra 355ml',
+                      '\$28.00',
+                      'imgs/Beerhouse/Bocanegra355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Bohemia Clara 355 ml',
+                      '\$13.00',
+                      'imgs/Beerhouse/BohemiaClara355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Bohemia Clara No retornable 355',
+                      '\$16.00',
+                      'imgs/Beerhouse/BohemiaClaraNoretornable355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Bohemia Obscura 355',
+                      '\$13.00',
+                      'imgs/Beerhouse/BohemiaObscura355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Bohemia Obscura No retornable 355',
+                      '\$16.00',
+                      'imgs/Beerhouse/BohemiaObscuraNoretornable355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Brooklyn 355ml',
+                      '\$44.00',
+                      'imgs/Beerhouse/Brooklyn355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'BudLight 355ml',
+                      '\$14.00',
+                      'imgs/Beerhouse/BudLight355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Carta Blanca 355',
+                      '\$16.00',
+                      'imgs/Beerhouse/CartaBlanca355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Carta Blanca 940ml',
+                      '\$24.00',
+                      'imgs/Beerhouse/CartaBlanca940ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Carta Blanca No retornable 355',
+                      '\$16.00',
+                      'imgs/Beerhouse/CartaBlancaNoretornable355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Cornona 355ml',
+                      '\$14.00',
+                      'imgs/Beerhouse/Cornona355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Corona 1.2L',
+                      '\$34.00',
+                      'imgs/Beerhouse/Corona12L.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Corona 940ml',
+                      '\$30.00',
+                      'imgs/Beerhouse/Corona940ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Corona Ampolleta',
+                      '\$10.00',
+                      'imgs/Beerhouse/CoronaAmpolleta.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Corona Light 1.2L',
+                      '\$34.00',
+                      'imgs/Beerhouse/CoronaLight12L.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Corona Light 355ml',
+                      '\$14.00',
+                      'imgs/Beerhouse/CoronaLight355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Goose Island 355ml',
+                      '\$49.00',
+                      'imgs/Beerhouse/GooseIsland355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Gran Modelo 1L',
+                      '\$35.00',
+                      'imgs/Beerhouse/GranModelo1L.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Grimbergen 330ml',
+                      '\$32.00',
+                      'imgs/Beerhouse/Grimbergen330ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Heineken 1L',
+                      '\$33.00',
+                      'imgs/Beerhouse/Heineken1L.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Heineken 250 ml',
+                      '\$12.00',
+                      'imgs/Beerhouse/Heineken250ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Heineken 355 ml',
+                      '\$17.00',
+                      'imgs/Beerhouse/Heineken355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Heineken No Retornable 355 ml',
+                      '\$17.00',
+                      'imgs/Beerhouse/HeinekenNoRetornable355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Hoegaarden 355ml',
+                      '\$29.00',
+                      'imgs/Beerhouse/Hoegaarden355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Indio 1.18L',
+                      '\$31.00',
+                      'imgs/Beerhouse/Indio118L.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Indio 355ml',
+                      '\$13.00',
+                      'imgs/Beerhouse/Indio355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Indio 940ml',
+                      '\$25.00',
+                      'imgs/Beerhouse/Indio940ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Indio ampolleta',
+                      '\$7.00',
+                      'imgs/Beerhouse/Indioampolleta.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Indio No Retornable 355ml',
+                      '\$13.00',
+                      'imgs/Beerhouse/IndioNoRetornable355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'JabalI 330ml',
+                      '\$33.00',
+                      'imgs/Beerhouse/JabalI330ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Lagunitas Ipa 355ml',
+                      '\$31.00',
+                      'imgs/Beerhouse/LagunitasIpa355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Leon 1.2L',
+                      '\$34.00',
+                      'imgs/Beerhouse/Leon12L.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Leon 355ml',
+                      '\$14.00',
+                      'imgs/Beerhouse/Leon355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Leon No retornable 325',
+                      '\$14.00',
+                      'imgs/Beerhouse/LeonNoretornable325.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Michelob Ultra 355ml',
+                      '\$17.00',
+                      'imgs/Beerhouse/MichelobUltra355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Modelo Ambar 355ml',
+                      '\$16.00',
+                      'imgs/Beerhouse/ModeloAmbar355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Modelo Especial 355ml',
+                      '\$16.00',
+                      'imgs/Beerhouse/ModeloEspecial355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Modelo Trigo 355ml',
+                      '\$16.00',
+                      'imgs/Beerhouse/ModeloTrigo355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Montejo 355ml',
+                      '\$16.00',
+                      'imgs/Beerhouse/Montejo355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Negra Modelo 355ml',
+                      '\$16.00',
+                      'imgs/Beerhouse/NegraModelo355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Noche Buena 355ml',
+                      '\$15.00',
+                      'imgs/Beerhouse/NocheBuena355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Pacifico 355',
+                      '\$14.00',
+                      'imgs/Beerhouse/Pacifico355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Pacifico Ampolleta',
+                      '\$10.00',
+                      'imgs/Beerhouse/PacificoAmpolleta.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Pacifico Light Ampolleta',
+                      '\$10.00',
+                      'imgs/Beerhouse/PacificoLightAmpolleta.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Quilmes 355ml',
+                      '\$25.00',
+                      'imgs/Beerhouse/Quilmes355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Sol 355',
+                      '\$13.00',
+                      'imgs/Beerhouse/Sol355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Sol 940ml',
+                      '\$25.00',
+                      'imgs/Beerhouse/Sol940ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Sol Ampolleta',
+                      '\$7.00',
+                      'imgs/Beerhouse/SolAmpolleta.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Sol No retornable 355ml',
+                      '\$13.00',
+                      'imgs/Beerhouse/SolNoretornable355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Stella Artois 310ml',
+                      '\$19.00',
+                      'imgs/Beerhouse/StellaArtois310ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Superior 355ml',
+                      '\$13.00',
+                      'imgs/Beerhouse/Superior355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Tecate Ambar 355ml',
+                      '\$10.00',
+                      'imgs/Beerhouse/TecateAmbar355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Tecate Ambar 940ml',
+                      '\$25.00',
+                      'imgs/Beerhouse/TecateAmbar940ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Tecate Light 1.18L',
+                      '\$31.00',
+                      'imgs/Beerhouse/TecateLight118L.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Tecate Light 355ml',
+                      '\$13.00',
+                      'imgs/Beerhouse/TecateLight355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Tecate Light No Retornable 355ml',
+                      '\$10.00',
+                      'imgs/Beerhouse/TecateLightNoRetornable355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Tecate Roja 1.18L',
+                      '\$31.00',
+                      'imgs/Beerhouse/TecateRoja118L.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Tecate Roja 355ml',
+                      '\$10.00',
+                      'imgs/Beerhouse/TecateRoja355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Tempus 355ml',
+                      '\$30.00',
+                      'imgs/Beerhouse/Tempus355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Tuborg 330ml',
+                      '\$16.00',
+                      'imgs/Beerhouse/Tuborg330ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Victoria 1.2L',
+                      '\$34.00',
+                      'imgs/Beerhouse/Victoria12L.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Victoria 355ml',
+                      '\$14.00',
+                      'imgs/Beerhouse/Victoria355ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Victoria 940ml',
+                      '\$30.00',
+                      'imgs/Beerhouse/Victoria940ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Victoria Ampolleta',
+                      '\$10.00',
+                      'imgs/Beerhouse/VictoriaAmpolleta.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'XX Lager 1.18L',
+                      '\$34.00',
+                      'imgs/Beerhouse/XXLager118L.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'XX Lager 355ml',
+                      '\$11.00',
+                      'imgs/Beerhouse/XXLager355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'XX Lager Ambar 1.18L',
+                      '\$34.00',
+                      'imgs/Beerhouse/XXLagerAmbar118L.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'XX Lager Ambar 355',
+                      '\$11.00',
+                      'imgs/Beerhouse/XXLagerAmbar355.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'XX lager Ampolleta',
+                      '\$8.00',
+                      'imgs/Beerhouse/XXlagerAmpolleta.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                ],
+              )),
+          SizedBox(height: 15.0)
+        ],
       ),
     );
   }
