@@ -1,11 +1,560 @@
 import 'package:flutter/material.dart';
+import 'package:blogapp/screens/product_detail.dart';
+import 'package:blogapp/screens/product_page.dart';
+import '../constant.dart';
 
 class CandysParty extends StatelessWidget {
+  ProductPage productpage = ProductPage();
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: Text('Hola Mundo'),
+    return Scaffold(
+      backgroundColor: Color(0xFFFCFAF8),
+      body: ListView(
+        children: <Widget>[
+          SizedBox(height: 15.0),
+          Container(
+              padding: EdgeInsets.only(right: 15.0),
+              width: MediaQuery.of(context).size.width - 30.0,
+              height: MediaQuery.of(context).size.height - 50.0,
+              child: GridView.count(
+                crossAxisCount: 2,
+                primary: false,
+                crossAxisSpacing: 10.0,
+                mainAxisSpacing: 15.0,
+                childAspectRatio: 0.8,
+                children: <Widget>[
+                  productpage.build(
+                      'Agua mineral ciel 1.75lts',
+                      '\$23.00',
+                      'imgs/Candysparty/Aguamineralciel175lt.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Agua mineral Perrier',
+                      '\$29.00',
+                      'imgs/Candysparty/AguaMineralPerrier.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Agua Natural 1lt',
+                      '\$8.00',
+                      'imgs/Candysparty/AguaNatural1lt.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Agua Natural 500ml',
+                      '\$5.00',
+                      'imgs/Candysparty/AguaNatural500ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Benson Hedges Gold Cajetilla',
+                      '\$63.00',
+                      'imgs/Candysparty/BensonHedgesGoldCajetillas.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Benson Hedges Mentolados Cajetilla',
+                      '\$63.00',
+                      'imgs/Candysparty/BensonHedgesMentolCajetillas.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Bolsa de Hielo',
+                      '\$27.00',
+                      'imgs/Candysparty/BolsadeHielo.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Boost 235ml',
+                      '\$25.00',
+                      'imgs/Candysparty/Boost235ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Botana Mixta',
+                      '\$29.00',
+                      'imgs/Candysparty/BotanaMixta.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Cacahuate Enchilado',
+                      '\$29.00',
+                      'imgs/Candysparty/CacahuateEnchilado.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Cacahuate Garapiñado 200gr',
+                      '\$N/A',
+                      'imgs/Candysparty/CacahuateGarapinado200gr.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Cacahuate Japones',
+                      '\$69',
+                      'imgs/Candysparty/CacahuateJapones.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Cacahuate japones 160gr',
+                      '\$20',
+                      'imgs/Candysparty/Cacahuatejapones160gr.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Cacahuate Salado sol',
+                      '\$69',
+                      'imgs/Candysparty/CacahuateSaladosol.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Cacahuate surtido 185gr',
+                      '\$20',
+                      'imgs/Candysparty/Cacahuatesurtido185gr.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Cacahuate Tipo Español con Ajo',
+                      '\$69',
+                      'imgs/Candysparty/CacahuateTipoEspanolconAjo.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Canada Dry 237ml',
+                      '\$N/A',
+                      'imgs/Candysparty/CanadaDry237ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Candy Camel Cajetilla',
+                      '\$63.00',
+                      'imgs/Candysparty/candycamelcajetilla.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Charritos chile y limón',
+                      '\$15.00',
+                      'imgs/Candysparty/Charritoschileylimon.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Clamato 2.5lts',
+                      '\$69.00',
+                      'imgs/Candysparty/Clamato25lts.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Coca Cola 1lt',
+                      '\$20.00',
+                      'imgs/Candysparty/CocaCola1lt.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Coca Cola 2.5lt',
+                      '\$36.00',
+                      'imgs/Candysparty/CocaCola25lt.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Coca Cola light 2lt',
+                      '\$31.00',
+                      'imgs/Candysparty/CocaColalight2lt.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Crunchi sol Cuadritos',
+                      '\$N/A',
+                      'imgs/Candysparty/CrunchisolCuadritos.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Crunchisol Donitas Enchiladas',
+                      '\$N/A',
+                      'imgs/Candysparty/CrunchisolDonitasEnchiladas.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Cueritosde Cerdo 750gr',
+                      '\$59.00',
+                      'imgs/Candysparty/CueritosdeCerdo750gr.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Fresca 2lt',
+                      '\$29.00',
+                      'imgs/Candysparty/Fresca2lt.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'frituras sabor queso sol 300gr',
+                      '\$20.00',
+                      'imgs/Candysparty/friturassaborquesosol300gr.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Garbanzo Enchilado',
+                      '\$69.00',
+                      'imgs/Candysparty/GarbanzoEnchilado.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Garbanzo enchilado 140gr',
+                      '\$20.00',
+                      'imgs/Candysparty/Garbanzoenchilado140gr.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Ginger Ale 237 ml',
+                      '\$9.00',
+                      'imgs/Candysparty/GingerAle237ml.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Haba Enchilada',
+                      '\$69.00',
+                      'imgs/Candysparty/HabaEnchilada.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Haba Enchilada 120gr',
+                      '\$20.00',
+                      'imgs/Candysparty/HabaEnchilada120gr.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Halls Negras',
+                      '\$10.00',
+                      'imgs/Candysparty/HallsNegras.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Marlboro gold cajetilla',
+                      '\$63.00',
+                      'imgs/Candysparty/Marlborogoldcajetilla.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Marlboro Ice Xpress',
+                      '\$63.00',
+                      'imgs/Candysparty/MarlboroIceXpress.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Marlboro Rojo cajetilla',
+                      '\$63.00',
+                      'imgs/Candysparty/marlboroRojocajetilla.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Maruchan',
+                      '\$15.00',
+                      'imgs/Candysparty/Maruchan.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Monster Energy',
+                      '\$42.00',
+                      'imgs/Candysparty/MonsterEnergy.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Mundet Prisco',
+                      '\$12.00',
+                      'imgs/Candysparty/MundetPrisco.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Nectar del Valle Naranja',
+                      '\$19.00',
+                      'imgs/Candysparty/NectardelValleNaranja.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Nectar del Valle Uva',
+                      '\$19.00',
+                      'imgs/Candysparty/NectardelValleUva.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Nectar del Valle Piña',
+                      '\$19.00',
+                      'imgs/Candysparty/NectarPinaDelValle.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Ocean Spray Cranberry',
+                      '\$26.00',
+                      'imgs/Candysparty/OceanSprayCranberry.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Ocean Spray Uva',
+                      '\$26.00',
+                      'imgs/Candysparty/OceanSprayUva.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Papa Adobada sol',
+                      '\$59.00',
+                      'imgs/Candysparty/PapaAdobadasol.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Papa con queso sol',
+                      '\$59.00',
+                      'imgs/Candysparty/Papaconquesosol.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Papa Natural de sal',
+                      '\$59.00',
+                      'imgs/Candysparty/PapaNaturaldelsal.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Peñafiel 1lt',
+                      '\$12.00',
+                      'imgs/Candysparty/Penafiel1lt.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Pepita frita',
+                      '\$99.00',
+                      'imgs/Candysparty/Pepitafrita.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Pepitarusa100gr',
+                      '\$20.00',
+                      'imgs/Candysparty/Pepitarusa100gr.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'RedBull Energy',
+                      '\$40.00',
+                      'imgs/Candysparty/RedBullEnergy.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'RedBull Sugar Free',
+                      '\$40.00',
+                      'imgs/Candysparty/RedBullSugarFree.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'RedBull Tropical',
+                      '\$40.00',
+                      'imgs/Candysparty/RedBullTropical.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Rockaleta',
+                      '\$5.00',
+                      'imgs/Candysparty/Rockaleta.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Sal La Fina',
+                      '\$15.00',
+                      'imgs/Candysparty/SalLaFina.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Salsa Valentina',
+                      '\$15.00',
+                      'imgs/Candysparty/SalsaValentina.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Sangría Señorial 1lt',
+                      '\$21.00',
+                      'imgs/Candysparty/SangriaSenorial1lt.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Sangrita Viuda de Sanchez',
+                      '\$71.00',
+                      'imgs/Candysparty/SangritaViudadeSanchez.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Sidral Mundet 2lt',
+                      '\$29.00',
+                      'imgs/Candysparty/SidralMundet2lt.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Skyy Blue',
+                      '\$22.00',
+                      'imgs/Candysparty/SkyyBlue.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Sangria Señorial 2lt',
+                      '\$30.00',
+                      'imgs/Candysparty/SnagriaSenorial2lt.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Sprite 2lt',
+                      '\$29.00',
+                      'imgs/Candysparty/Sprite2lt.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Trident XtraCare Menta Tabletas',
+                      '\$10.00',
+                      'imgs/Candysparty/Trident.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Tubi Charro Naturales',
+                      '\$N/A',
+                      'imgs/Candysparty/TubiCharroNaturales.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                  productpage.build(
+                      'Tupsi Pop',
+                      '\$5',
+                      'imgs/Candysparty/Tupsi.png',
+                      false,
+                      false,
+                      "Detalle de Producto",
+                      context),
+                ],
+              )),
+          SizedBox(height: 15.0)
+        ],
       ),
     );
   }
