@@ -1,11 +1,810 @@
 import 'package:flutter/material.dart';
+import 'package:blogapp/screens/product_detail.dart';
+import 'package:blogapp/screens/product_page.dart';
+import '../constant.dart';
 
 class HardDrinks extends StatelessWidget {
+  ProductPage productpage = ProductPage();
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: Text('Hola Mundo'),
+    return Scaffold(
+      backgroundColor: Color(0xFFFCFAF8),
+      body: ListView(
+        children: <Widget>[
+          SizedBox(height: 15.0),
+          Container(
+              padding: EdgeInsets.only(right: 15.0),
+              width: MediaQuery.of(context).size.width - 30.0,
+              height: MediaQuery.of(context).size.height - 50.0,
+              child: GridView.count(
+                crossAxisCount: 2,
+                primary: false,
+                crossAxisSpacing: 10.0,
+                mainAxisSpacing: 15.0,
+                childAspectRatio: 0.8,
+                children: <Widget>[
+                  productpage.build(
+                      '100 años 700 ml',
+                      '\$140.00',
+                      'imgs/HardDrink/100anosreposado.png',
+                      false,
+                      false,
+                      "Detalle 100 años",
+                      context),
+                  productpage.build(
+                      '400 Conejo',
+                      '\$435.00',
+                      'imgs/HardDrink/400conejos.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      '1800 Cristalino',
+                      '\$635.00',
+                      'imgs/HardDrink/1800cristalino.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Absinth',
+                      '\$555.00',
+                      'imgs/HardDrink/Absinth.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Absolut Azul',
+                      '\$259.00',
+                      'imgs/HardDrink/AbsolutAzul.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Absolut Mandarina',
+                      '\$259.00',
+                      'imgs/HardDrink/AbsolutMandrin.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Alacrán Jóven ',
+                      '\$639.00',
+                      'imgs/HardDrink/AlacranJoven.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Alipús San Juan del Río',
+                      '\$717.00',
+                      'imgs/HardDrink/AlipusSanJuandelRio.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Appleton Estate',
+                      '\$250.00',
+                      'imgs/HardDrink/AppletonEstate.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Appleton Special',
+                      '\$635.00',
+                      'imgs/HardDrink/AppletonSpecial.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Azteca de Oro',
+                      '\$150.00',
+                      'imgs/HardDrink/AztecadeOro.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Bacardí Blanco',
+                      '\$165.00',
+                      'imgs/HardDrink/BacardiBlanco.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Bacardí Blanco 980 ml',
+                      '\$210.00',
+                      'imgs/HardDrink/BacardiBlanco980ml.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Bacardí Blanco 1750',
+                      '\$310.00',
+                      'imgs/HardDrink/BacardiBlanco1750.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Bacardi Solera',
+                      '\$220.00',
+                      'imgs/HardDrink/BacardiSolera.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Baileys Original',
+                      '\$315.00',
+                      'imgs/HardDrink/BaileysOriginal.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Beefeater',
+                      '\$412.00',
+                      'imgs/HardDrink/Beefeater.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Beringer White Zinfandel',
+                      '\$201.00',
+                      'imgs/HardDrink/BeringerWhiteZinfandel.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Black & White',
+                      '\$200.00',
+                      'imgs/HardDrink/BlackWhite.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Bombay',
+                      '\$475.00',
+                      'imgs/HardDrink/Bombay.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Bruxo',
+                      '\$515.00',
+                      'imgs/HardDrink/Bruxo.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Buchanans 12 años',
+                      '\$780.00',
+                      'imgs/HardDrink/Buchanans12anos.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Buchanans 18 años',
+                      '\$780.00',
+                      'imgs/HardDrink/Buchanans18anos.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Buchanans Master',
+                      '\$895.00',
+                      'imgs/HardDrink/BuchanansMaster.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Bulldog',
+                      '\$550.00',
+                      'imgs/HardDrink/Bulldog.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Captain Morgan Spiced 750 ml',
+                      '\$149.00',
+                      'imgs/HardDrink/CaptainMorganSpiced750ml.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Cardenal de Mendoza',
+                      '\$1063.00',
+                      'imgs/HardDrink/Cardenaldeendoza.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Cazadores Reposado',
+                      '\$230.00',
+                      'imgs/HardDrink/CazadoresReposado.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Centenario Plata 700 ml',
+                      '\$305.00',
+                      'imgs/HardDrink/CentenarioPlata700ml.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Centenario Plata 950 ml',
+                      '\$317.00',
+                      'imgs/HardDrink/CentenarioPlata950ml.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Centenario Reposado',
+                      '\$255.00',
+                      'imgs/HardDrink/centenarioreposado.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Chilensis Reserva',
+                      '\$180.00',
+                      'imgs/HardDrink/ChilensisReserva.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Chivas Regal 12 años',
+                      '\$695.00',
+                      'imgs/HardDrink/ChivasRegal12anos.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Courvoisier V.S.',
+                      '\$625.00',
+                      'imgs/HardDrink/COURVOISIERVS.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Cune de Rioja Semidulce',
+                      '\$200.00',
+                      'imgs/HardDrink/CuneRiojaSemidulce.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Don julio 70',
+                      '\$570.00',
+                      'imgs/HardDrink/Donjulio70.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Don Julio 70 Añejo Cristalino',
+                      '\$835.00',
+                      'imgs/HardDrink/DonJulioReposado.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Don Pedro',
+                      '\$160.00',
+                      'imgs/HardDrink/DonPedro.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'El Jimador reposado',
+                      '\$195.00',
+                      'imgs/HardDrink/ElJimadorreposado.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Finca las Moras Malbec',
+                      '\$142.00',
+                      'imgs/HardDrink/FincalasMorasMalbec.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Frangelico',
+                      '\$350.00',
+                      'imgs/HardDrink/Frangelico.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Fundador',
+                      '\$185.00',
+                      'imgs/HardDrink/Fundador.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Glenfiddich 12 años',
+                      '\$915.00',
+                      'imgs/HardDrink/Glenfiddich12anos.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Grey Goose',
+                      '\$660.00',
+                      'imgs/HardDrink/GreyGoose.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Havana Club 750 ml',
+                      '\$442.00',
+                      'imgs/HardDrink/HavanaClub750ml.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Hendricks',
+                      '\$750.00',
+                      'imgs/HardDrink/Hendricks.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Hennessy V.S.O.P',
+                      '\$1080.00',
+                      'imgs/HardDrink/HennessyVSOP.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Herradura Reposado 700',
+                      '\$549.00',
+                      'imgs/HardDrink/HerraduraReposado700.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Hornitos Reposado 700',
+                      '\$275.00',
+                      'imgs/HardDrink/HornitosReposado700.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Hpnotiq',
+                      '\$650.00',
+                      'imgs/HardDrink/Hpnotiq.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Jack Daniels',
+                      '\$415.00',
+                      'imgs/HardDrink/JackDaniels.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Jagermeifter',
+                      '\$320.00',
+                      'imgs/HardDrink/Jagermeifter.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Jameson',
+                      '\$371.00',
+                      'imgs/HardDrink/Jameson.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build('J B', '\$240.00', 'imgs/HardDrink/JB.png',
+                      false, false, "Detalle producto", context),
+                  productpage.build(
+                      'Johnnie Walker Etiqueta Negra',
+                      '\$830.00',
+                      'imgs/HardDrink/JohnnieWalkerEtiquetaNegra.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Johnnie Walker Etiqueta Roja',
+                      '\$325.00',
+                      'imgs/HardDrink/JohnnieWalkerEtiquetaRoja.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Jose Cuervo Especial',
+                      '\$205.00',
+                      'imgs/HardDrink/JoseCuervoEspecial.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Jose Cuervo tradicional',
+                      '\$330.00',
+                      'imgs/HardDrink/JoseCuervotradicional.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Kahlua Licor de Cafe',
+                      '\$166.00',
+                      'imgs/HardDrink/KahluaLicordeCafe.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Karat 1 lt',
+                      '\$130.00',
+                      'imgs/HardDrink/Karat1lt.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Kraken Black Spiced',
+                      '\$165.00',
+                      'imgs/HardDrink/KrakenBlackSpiced.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'L.A. Cetto Cabernet Sauvignon',
+                      '\$136.00',
+                      'imgs/HardDrink/LACetto.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'L.A. Cetto Blanc Zinfandel',
+                      '\$131.00',
+                      'imgs/HardDrink/LACettoBlancdeZinfandel.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Licor 43',
+                      '\$430.00',
+                      'imgs/HardDrink/Licor43.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Macallan Doble Cask',
+                      '\$1260.00',
+                      'imgs/HardDrink/MacallanDobleCask.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Maestro Dobel',
+                      '\$650.00',
+                      'imgs/HardDrink/MaestroDobel.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Malibu',
+                      '\$187.00',
+                      'imgs/HardDrink/Malibu.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Martell V.S.',
+                      '\$700.00',
+                      'imgs/HardDrink/Martellvs.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Martell V.S.O.P',
+                      '\$949.00',
+                      'imgs/HardDrink/MartellVSOP.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Matusalem Clasico',
+                      '\$224.00',
+                      'imgs/HardDrink/MatusalemClasico.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Matusalem Platino',
+                      '\$190.00',
+                      'imgs/HardDrink/MatusalemPlatino.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Mezcal Union Uno',
+                      '\$N/A.00',
+                      'imgs/HardDrink/MezcalUnionUno.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Midori',
+                      '\$405.00',
+                      'imgs/HardDrink/Midori.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Moet Chandon Brut',
+                      '\$1050.00',
+                      'imgs/HardDrink/MoetChandonBrut.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Moet Chandon Ice Imperial',
+                      '\$1655.00',
+                      'imgs/HardDrink/MoetChandonIceImperial.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Moet Chandon Rose',
+                      '\$1295.00',
+                      'imgs/HardDrink/MoetChandonRose.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Monte Xanic Merlot',
+                      '\$545.00',
+                      'imgs/HardDrink/MonteXanicMerlot.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Nuvo',
+                      '\$770.00',
+                      'imgs/HardDrink/Nuvo.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Oso Negro Ginebra',
+                      '\$120.00',
+                      'imgs/HardDrink/OsoNegroGinebra.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Oso Ngero 1 lt',
+                      '\$90.00',
+                      'imgs/HardDrink/OsoNgero1lt.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Passport Scotch',
+                      '\$160.00',
+                      'imgs/HardDrink/PassportScotch.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Remy Martin 1738',
+                      '\$1420.00',
+                      'imgs/HardDrink/RemyMartin1738.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Remy Martin V.S.O.P.',
+                      '\$1100.00',
+                      'imgs/HardDrink/RemyMartinVSOP.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Riunite Lambrusco',
+                      '\$140.00',
+                      'imgs/HardDrink/RiuniteLambrusco.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Riunite Prosecco',
+                      '\$280.00',
+                      'imgs/HardDrink/RiuniteProsecco.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Sauza',
+                      '\$229.00',
+                      'imgs/HardDrink/Sauza.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Skyy Vodka',
+                      '\$165.00',
+                      'imgs/HardDrink/SkyyVodka.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Smirnoff 750 ml',
+                      '\$185.00',
+                      'imgs/HardDrink/Smirnoff750ml.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Smirnoff Tamarindo',
+                      '\$200.00',
+                      'imgs/HardDrink/SmirnoffTamarindo.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Stolichnaya',
+                      '\$265.00',
+                      'imgs/HardDrink/Stolichnaya.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Tanqueray',
+                      '\$455.00',
+                      'imgs/HardDrink/Tanqueray.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Terry Centenario',
+                      '\$280.00',
+                      'imgs/HardDrink/TerryCentenario.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Torres 10',
+                      '\$290.00',
+                      'imgs/HardDrink/Torres10.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Torres 15',
+                      '\$510.00',
+                      'imgs/HardDrink/Torres15.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Torres 20',
+                      '\$875.00',
+                      'imgs/HardDrink/Torres20.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Wyborowa',
+                      '\$139.00',
+                      'imgs/HardDrink/Wyborowa.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'XA DOMECQ',
+                      '\$120.00',
+                      'imgs/HardDrink/XADOMECQ.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                  productpage.build(
+                      'Zacapa 23',
+                      '\$1100.00',
+                      'imgs/HardDrink/Zacapa23.png',
+                      false,
+                      false,
+                      "Detalle producto",
+                      context),
+                ],
+              )),
+          SizedBox(height: 15.0)
+        ],
       ),
     );
   }
