@@ -1,11 +1,548 @@
 import 'package:flutter/material.dart';
+import 'package:blogapp/screens/product_detail.dart';
+import 'package:blogapp/screens/product_page.dart';
+import '../constant.dart';
 
+// ignore: must_be_immutable
 class PistosMix extends StatelessWidget {
+/*   var pistoxmixarr = [
+    {
+      {
+        'Alfa Romeo',
+        '95',
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Aston Martin',
+        '129',
+        true,
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Bentley Mango o Tamarindo',
+        '109',
+        true,
+        true,
+        true,
+        true,
+        false,
+        true,
+        false,
+        false,
+        false,
+      },
+      {
+        'Bugatti',
+        '85',
+        true,
+        false,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Cadillac',
+        '109',
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Citroen',
+        '85',
+        true,
+        false,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Corvette',
+        '119',
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Ferrari',
+        '109',
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Hennessey',
+        '129',
+        true,
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Lamborghini',
+        '99',
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Lexus',
+        '129',
+        true,
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Likan',
+        '129',
+        true,
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Maserati',
+        '99',
+        true,
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Mastretta',
+        '139',
+        true,
+        true,
+        true,
+        true,
+        true,
+        false,
+        true,
+        false,
+        false,
+      },
+      {
+        'Mclaren',
+        '109',
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        true,
+        false,
+      },
+      {
+        'MG',
+        '109',
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Nos F1',
+        '139',
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Rolls Royce',
+        '119',
+        true,
+        false,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Saleen ',
+        '119',
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Porsche',
+        '109',
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        true,
+      },
+      {
+        'Venom',
+        '129',
+        true,
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+      },
+      {
+        'Viper',
+        '109',
+        true,
+        true,
+        true,
+        true,
+        false,
+        false,
+        false,
+        false,
+        false,
+      },
+    }
+  ];
+  var cervesas = [
+    'Barrilito',
+    'Blanc',
+    'Bocanegra',
+    'Bohemia Clara',
+    'Bohemia Clásica',
+    'Bohemia Obscura',
+    'Bohemina Weizen ',
+    'Brooklyn',
+    'Bud Light',
+    'Budweiser',
+    'Carta Blanca',
+    'Corona',
+    'Corona l',
+    'Cucapa',
+    'Goose Island',
+    'Gran Modelo',
+    'Grimbergen',
+    'Heineken',
+    'Hoegaarden',
+    'Indio ',
+    'Jabalí',
+    'Lagunitas Ipa',
+    'León',
+    'Michelob Ultra',
+    'Modelo Ambar ',
+    'Modelo Especial',
+    'Modelo Trigo',
+    'Montejo',
+    'Negra Modelo',
+    'Noche Buena ',
+    'Pacifico',
+    'Pacifico l',
+    'Quilmes',
+    'Sol',
+    'Stella Artois',
+    'Superior',
+    'Tecate Ambar',
+    'Tecate Ambar ',
+    'Tecate Light',
+    'Tecate Roja',
+    'Tempus',
+    'Tuborg',
+    'Victoria',
+    'XX Ambar',
+    'xx Lager',
+  ];
+ */
+
+  ProductPage productpage = ProductPage();
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: Text('Hola Mundo'),
+    return Scaffold(
+      backgroundColor: Color(0xFFFCFAF8),
+      body: ListView(
+        children: <Widget>[
+          SizedBox(height: 15.0),
+          Container(
+              padding: EdgeInsets.only(right: 15.0),
+              width: MediaQuery.of(context).size.width - 30.0,
+              height: MediaQuery.of(context).size.height - 50.0,
+              child: GridView.count(
+                crossAxisCount: 2,
+                primary: false,
+                crossAxisSpacing: 10.0,
+                mainAxisSpacing: 15.0,
+                childAspectRatio: 0.8,
+                children: <Widget>[
+                  productpage.build(
+                      'Alfa romeo',
+                      '\$95.00',
+                      'imgs/PistosMix/Alfaromeo.png',
+                      false,
+                      false,
+                      "Michelada tradicional sal, limón y  escarchado",
+                      context),
+                  productpage.build(
+                      'Aston Martin',
+                      '\$129.00',
+                      'imgs/PistosMix/astonmartin.png',
+                      false,
+                      false,
+                      "Michelada sabor Bluberry con Red Bull",
+                      context),
+                  productpage.build(
+                      'Bentley',
+                      '\$109.00',
+                      'imgs/PistosMix/Bentley.png',
+                      false,
+                      false,
+                      "Cerveza sabor a mango con nieve del mismo sabor",
+                      context),
+                  productpage.build(
+                      'Bugatti',
+                      '\$85.00',
+                      'imgs/PistosMix/Bugatti.png',
+                      false,
+                      false,
+                      "Sangría preparada con limón y sal ",
+                      context),
+                  productpage.build(
+                      'Cadillac',
+                      '\$109.00',
+                      'imgs/PistosMix/Cadillac.png',
+                      false,
+                      false,
+                      "Chabelita cerveza mezclada con refresco rojo, sal Y limón",
+                      context),
+                  productpage.build(
+                      'Citroen',
+                      '\$85.00',
+                      'imgs/PistosMix/Citroen.png',
+                      false,
+                      false,
+                      "Tehuacan preparado con limón y sal",
+                      context),
+                  productpage.build(
+                      'Corvette',
+                      '\$119.00',
+                      'imgs/PistosMix/Corvette.png',
+                      false,
+                      false,
+                      "Cerveza con sidra Strongbow gold sabor a frambuesa",
+                      context),
+                  productpage.build(
+                      'Ferrari',
+                      '\$109.00',
+                      'imgs/PistosMix/Ferrari.png',
+                      false,
+                      false,
+                      "Mezcla de sabor cerveza y chamoy",
+                      context),
+                  productpage.build(
+                      'Hennessey',
+                      '\$129.00',
+                      'imgs/PistosMix/Hennessey.png',
+                      false,
+                      false,
+                      "Michelada sabor frutos rojos con Red Bull",
+                      context),
+                  productpage.build(
+                      'Lamborghinni',
+                      '\$99.00',
+                      'imgs/PistosMix/Lamborghinni.png',
+                      false,
+                      false,
+                      "Michelada tradicional sal, limón y  escarchado",
+                      context),
+                  productpage.build(
+                      'Lexus',
+                      '\$129.00',
+                      'imgs/PistosMix/Lexus.png',
+                      false,
+                      false,
+                      "Michelada sabor manzana verde con Red Bull",
+                      context),
+                  productpage.build(
+                      'Likan',
+                      '\$129.00',
+                      'imgs/PistosMix/Likan.png',
+                      false,
+                      false,
+                      "Michelada sabor maracuya-chamoy verde con Red Bull",
+                      context),
+                  productpage.build(
+                      'Masserati',
+                      '\$99.00',
+                      'imgs/PistosMix/Masserati.png',
+                      false,
+                      false,
+                      "Michelada Cubana, salsa inglesa, salsa Maggy, salsa Valentina, salsa Tabasco, limón y sal",
+                      context),
+                  productpage.build(
+                      'Mastretta',
+                      '\$139.00',
+                      'imgs/PistosMix/Mastretta.png',
+                      false,
+                      false,
+                      "Michelada sabor pepino-morita y una onza de Mezcal ",
+                      context),
+                  productpage.build(
+                      'Mclaren',
+                      '\$109.00',
+                      'imgs/PistosMix/Mclaren.png',
+                      false,
+                      false,
+                      "Clamachela, cerveza con clamato natural",
+                      context),
+                  productpage.build(
+                      'Mg',
+                      '\$109.00',
+                      'imgs/PistosMix/Mg.png',
+                      false,
+                      false,
+                      "Gomichela, mezcla de chamoy y gomitas ",
+                      context),
+                  productpage.build(
+                      'NosF1',
+                      '\$139.00',
+                      'imgs/PistosMix/NosF1.png',
+                      false,
+                      false,
+                      "Cerveza sabor a uva y jugo de litchi",
+                      context),
+                  productpage.build(
+                      'Porshe',
+                      '\$109.00',
+                      'imgs/PistosMix/porshe.png',
+                      false,
+                      false,
+                      "Cerveza con sabores, fresa, coco, maracuya, piña, uva, durazno y chamoy",
+                      context),
+                  productpage.build(
+                      'RollRoyce',
+                      '\$119.00',
+                      'imgs/PistosMix/RollRoyce.png',
+                      false,
+                      false,
+                      "Calimocho, Vino tinto con limón, sal y Coca cola",
+                      context),
+                  productpage.build(
+                      'Saleen',
+                      '\$119.00',
+                      'imgs/PistosMix/Saleen.png',
+                      false,
+                      false,
+                      "Cerveza con Skyy Blue y sabor a mandarina",
+                      context),
+                  productpage.build(
+                      'Venom',
+                      '\$129.00',
+                      'imgs/PistosMix/Venom.png',
+                      false,
+                      false,
+                      "Michelada sabor piña colada con Red Bull",
+                      context),
+                  productpage.build(
+                      'Viper',
+                      '\$109.00',
+                      'imgs/PistosMix/Viper.png',
+                      false,
+                      false,
+                      "Cerveza con sabor a fresa- kiwi acompañada de bing picosito",
+                      context),
+                ],
+              )),
+          SizedBox(height: 15.0)
+        ],
       ),
     );
   }
